Create projectCard component
